visualizer/app/requirements.txt
flask
azure-cosmos
azure-identity
matplotlib